+++ LibCarla/source/carla/client/detail/ActorFactory.cpp
@@ -11,6 +11,9 @@
 #include "carla/client/Actor.h"
 #include "carla/client/LaneInvasionSensor.h"
 #include "carla/client/ServerSideSensor.h"
+#ifdef RSS_ENABLED
+#include "carla/rss/RssSensor.h"
+#endif
 #include "carla/client/TrafficLight.h"
 #include "carla/client/TrafficSign.h"
 #include "carla/client/Vehicle.h"
@@ -74,6 +77,10 @@ namespace detail {
     auto init = ActorInitializer{description, episode};
     if (description.description.id == "sensor.other.lane_invasion") {
       return MakeActorImpl<LaneInvasionSensor>(std::move(init), gc);
+#ifdef RSS_ENABLED
+    } else if (description.description.id == "sensor.other.rss") {
+      return MakeActorImpl<RssSensor>(std::move(init), gc);
+#endif
     } else if (description.HasAStream()) {
       return MakeActorImpl<ServerSideSensor>(std::move(init), gc);
     } else if (StringUtil::StartsWith(description.description.id, "vehicle.")) {