+++ PythonAPI/carla/agents/navigation/basic_agent.py
@@ -122,3 +122,10 @@ class BasicAgent(Agent):
             control = self._local_planner.run_step(debug=debug)
 
         return control
+
+    def done(self):
+        """
+        Check whether the agent has reached its destination.
+        :return bool
+        """
+        return self._local_planner.done()